In: #6 (REFACTOR) Move the file 'ListNurseScreen.kt' and 'Login.kt' to 'ui' package. - Simplified imports by using '*', reducing redundancy in the affected packages.
