+++ app/src/main/java/com/example/hospitalfrontend/MainActivity.kt
@@ -4,25 +4,14 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxHeight
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Button
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
+import androidx.compose.foundation.layout.*
+import androidx.compose.material3.*
+import androidx.compose.runtime.*
 import androidx.compose.runtime.saveable.rememberSaveable
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import com.example.hospitalfrontend.ui.*
 import com.example.hospitalfrontend.ui.theme.HospitalFrontEndTheme
 
 class MainActivity : ComponentActivity() {
@@ -61,6 +50,7 @@ fun MyAppHomePage() {
             onScreenSelected = { selectedScreen -> nextScreen = selectedScreen })
     }
 }
+
 @Composable
 fun BottomNavigationBar(
     currentScreen: String, onScreenSelected: (String) -> Unit, modifier: Modifier = Modifier
